# api/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Order APIs",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get Orders",
                "operationId": "ApiV1GetOrders",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "[{'id': 10,'total_distance': 170201,'status': 'UNASSIGNED'},{'id': 11,'total_distance': 170211,'status':'UNASSIGNED'}]"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new Order",
                "operationId": "ApiV1saveOrder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "origin": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    },
                                    "example": [
                                        "28.053049",
                                        "76.107712"
                                    ]
                                },
                                "destination": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    },
                                    "example": [
                                        "26.912434",
                                        "75.787270"
                                    ]
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{'total_distance': 170201,'status': 'UNASSIGNED','id': 8}"
                    },
                    "422": {
                        "description": "{'error': 'Invalid location, please enter valid location or check google api'}"
                    }
                }
            }
        },
        "/orders/{id}": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update order status",
                "operationId": "ApiV1pdateUOrderStatus",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "TAKEN"
                                }
                            }
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{'status': 'SUCCESS'}"
                    },
                    "422": {
                        "description": "{'error': 'Invalid location, please enter valid location or check google api'}"
                    }
                }
            }
        }
    },
    "definitions": {}
}